# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "ORKG-NLP-API",
        "version": "0.3.1"
    },
    "servers": [
        {
            "url": "",
            "description": ""
        }
    ],
    "paths": {
        "/clustering/predicates": {
            "post": {
                "tags": [
                    "clustering"
                ],
                "summary": "Recommends Predicates",
                "operationId": "recommends_predicates_clustering_predicates_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PredicatesClusteringRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PredicatesClusteringResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clustering/bioassays": {
            "post": {
                "tags": [
                    "clustering"
                ],
                "summary": "Semantifies Bioassays",
                "operationId": "semantifies_bioassays_clustering_bioassays_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BioAssaysClusteringRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BioAssaysClusteringResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/annotation/csner": {
            "post": {
                "tags": [
                    "annotation"
                ],
                "summary": "Annotates Cs Paper",
                "operationId": "annotates_cs_paper_annotation_csner_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CSNerAnnotationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSNerAnnotationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/annotation/agriner": {
            "post": {
                "tags": [
                    "annotation"
                ],
                "summary": "Annotates Agri Paper",
                "operationId": "annotates_agri_paper_annotation_agriner_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgriNerAnnotationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgriNerAnnotationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tools/text/summarize": {
            "post": {
                "tags": [
                    "text"
                ],
                "summary": "Summarizes Text",
                "operationId": "summarizes_text_tools_text_summarize_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SummarizeTextRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SummarizeTextResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tools/text/classify": {
            "post": {
                "tags": [
                    "text"
                ],
                "summary": "Classifies Sentence",
                "operationId": "classifies_sentence_tools_text_classify_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassifySentenceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassifySentenceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AgriNerAnnotationRequest": {
                "title": "AgriNerAnnotationRequest",
                "required": [
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    }
                }
            },
            "AgriNerAnnotationResponse": {
                "title": "AgriNerAnnotationResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__annotation__AgriNerAnnotationResponse__Payload"
                    }
                }
            },
            "Annotation": {
                "title": "Annotation",
                "required": [
                    "concept",
                    "entities"
                ],
                "type": "object",
                "properties": {
                    "concept": {
                        "title": "Concept",
                        "type": "string"
                    },
                    "entities": {
                        "title": "Entities",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "BioAssayLabel": {
                "title": "BioAssayLabel",
                "required": [
                    "property",
                    "resources"
                ],
                "type": "object",
                "properties": {
                    "property": {
                        "$ref": "#/components/schemas/Predicate"
                    },
                    "resources": {
                        "title": "Resources",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Resource"
                        }
                    }
                }
            },
            "BioAssaysClusteringRequest": {
                "title": "BioAssaysClusteringRequest",
                "required": [
                    "text"
                ],
                "type": "object",
                "properties": {
                    "text": {
                        "title": "Text",
                        "type": "string"
                    }
                }
            },
            "BioAssaysClusteringResponse": {
                "title": "BioAssaysClusteringResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__clustering__BioAssaysClusteringResponse__Payload"
                    }
                }
            },
...
            "CSNerAnnotationRequest": {
                "title": "CSNerAnnotationRequest",
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "abstract": {
                        "title": "Abstract",
                        "type": "string"
                    }
                }
            },
            "CSNerAnnotationResponse": {
                "title": "CSNerAnnotationResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__annotation__CSNerAnnotationResponse__Payload"
                    }
                }
            },
            "ClassifySentenceRequest": {
                "title": "ClassifySentenceRequest",
                "required": [
                    "sentence",
                    "labels"
                ],
                "type": "object",
                "properties": {
                    "sentence": {
                        "title": "Sentence",
                        "type": "string"
                    },
                    "labels": {
                        "title": "Labels",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ClassifySentenceResponse": {
                "title": "ClassifySentenceResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__text__ClassifySentenceResponse__Payload"
                    }
                }
            },
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "Predicate": {
                "title": "Predicate",
                "required": [
                    "id",
                    "label"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "label": {
                        "title": "Label",
                        "type": "string"
                    }
                }
            },
            "PredicatesClusteringRequest": {
                "title": "PredicatesClusteringRequest",
                "required": [
                    "title",
                    "abstract"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "abstract": {
                        "title": "Abstract",
                        "type": "string"
                    }
                }
            },
            "PredicatesClusteringResponse": {
                "title": "PredicatesClusteringResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__clustering__PredicatesClusteringResponse__Payload"
                    }
                }
            },
            "Resource": {
                "title": "Resource",
                "required": [
                    "id",
                    "label"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "label": {
                        "title": "Label",
                        "type": "string"
                    }
                }
            },
...
            "SummarizeTextRequest": {
                "title": "SummarizeTextRequest",
                "required": [
                    "text",
                    "ratio"
                ],
                "type": "object",
                "properties": {
                    "text": {
                        "title": "Text",
                        "type": "string"
                    },
                    "ratio": {
                        "title": "Ratio",
                        "type": "number"
                    }
                }
            },
            "SummarizeTextResponse": {
                "title": "SummarizeTextResponse",
                "required": [
                    "timestamp",
                    "uuid",
                    "payload"
                ],
                "type": "object",
                "properties": {
                    "timestamp": {
                        "title": "Timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string",
                        "format": "uuid"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/app__models__text__SummarizeTextResponse__Payload"
                    }
                }
            },
...
            "TitleAbstractAnnotations": {
                "title": "TitleAbstractAnnotations",
                "required": [
                    "title",
                    "abstract"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Annotation"
                        }
                    },
                    "abstract": {
                        "title": "Abstract",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Annotation"
                        }
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
            "app__models__annotation__AgriNerAnnotationResponse__Payload": {
                "title": "Payload",
                "required": [
                    "annotations"
                ],
                "type": "object",
                "properties": {
                    "annotations": {
                        "title": "Annotations",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Annotation"
                        }
                    }
                }
            },
            "app__models__annotation__CSNerAnnotationResponse__Payload": {
                "title": "Payload",
                "required": [
                    "annotations"
                ],
                "type": "object",
                "properties": {
                    "annotations": {
                        "$ref": "#/components/schemas/TitleAbstractAnnotations"
                    }
                }
            },
            "app__models__clustering__BioAssaysClusteringResponse__Payload": {
                "title": "Payload",
                "required": [
                    "labels"
                ],
                "type": "object",
                "properties": {
                    "labels": {
                        "title": "Labels",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BioAssayLabel"
                        }
                    }
                }
            },
            "app__models__clustering__PredicatesClusteringResponse__Payload": {
                "title": "Payload",
                "required": [
                    "predicates"
                ],
                "type": "object",
                "properties": {
                    "predicates": {
                        "title": "Predicates",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Predicate"
                        }
                    }
                }
            },
...
            "app__models__text__ClassifySentenceResponse__Payload": {
                "title": "Payload",
                "required": [
                    "labels",
                    "scores",
                    "sequence"
                ],
                "type": "object",
                "properties": {
                    "labels": {
                        "title": "Labels",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "scores": {
                        "title": "Scores",
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    },
                    "sequence": {
                        "title": "Sequence",
                        "type": "string"
                    }
                }
            },
            "app__models__text__SummarizeTextResponse__Payload": {
                "title": "Payload",
                "required": [
                    "summary"
                ],
                "type": "object",
                "properties": {
                    "summary": {
                        "title": "Summary",
                        "type": "string"
                    }
                }
            }
        }
    }
}